+++ src/output.ts
@@ -23,15 +23,28 @@ interface OutputTestInterface {
   output: string | null
   // eslint-disable-next-line @typescript-eslint/naming-convention
   test_code: string
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  task_id?: number
 }
 
-const OUTPUT_VERSION = 2
+type ExerciseConfig = {
+  custom?: {
+    'version.tests.compatibility'?: 'jest-27'
+    'flag.tests.task-per-describe': boolean
+    'flag.tests.may-run-long': boolean
+    'flag.tests.includes-optional': boolean
+  }
+}
+
+const OUTPUT_VERSION = 3
+
 export class Output {
   private results: Partial<OutputInterface> & Pick<OutputInterface, 'tests'>
   private readonly globalConfig: Config.GlobalConfig
   private readonly outputFile: string
   private readonly sources: Record<string, ParsedSource>
   private readonly tests: Record<string, ReturnType<typeof extractTests>>
+  private readonly config: null | ExerciseConfig
 
   constructor(globalConfig: Config.GlobalConfig) {
     this.globalConfig = globalConfig
@@ -39,6 +52,7 @@ export class Output {
     this.outputFile =
       this.globalConfig.outputFile ?? path.join(process.cwd(), 'results.json')
 
+    this.config = findConfig(path.dirname(this.outputFile))
     this.sources = {}
     this.tests = {}
   }
@@ -131,8 +145,22 @@ export class Output {
         return { ...test, test_code: null }
       }
 
-      // eslint-disable-next-line @typescript-eslint/naming-convention
-      return { ...test, test_code: testCase.testCode(parsedSource.source) }
+      // Version 3 optionally can output the task ID
+      if (this.configFlag('flag.tests.task-per-describe')) {
+        return {
+          ...test,
+          // eslint-disable-next-line @typescript-eslint/naming-convention
+          task_id: testCase.topLevelIndex,
+          // eslint-disable-next-line @typescript-eslint/naming-convention
+          test_code: testCase.testCode(parsedSource.source),
+        }
+      }
+
+      return {
+        ...test,
+        // eslint-disable-next-line @typescript-eslint/naming-convention
+        test_code: testCase.testCode(parsedSource.source),
+      }
     })
 
     // Re-order the output so that tests output shows below main output
@@ -144,6 +172,16 @@ export class Output {
       2
     )
     fs.writeFileSync(this.outputFile, artifact)
+  }
+
+  private configFlag(
+    flag: keyof NonNullable<ExerciseConfig['custom']>
+  ): boolean {
+    if (!this.config || !this.config.custom) {
+      return false
+    }
+
+    return this.config.custom[flag] as boolean
   }
 
   public testFinished(
@@ -389,4 +427,23 @@ function sanitizeErrorMessage(specFilePath: string, message: string): string {
   }
 
   return message
+}
+
+function findConfig(rootDir: string): null | ExerciseConfig {
+  const configPaths = [
+    path.join(rootDir, '.meta', 'config.json'),
+    path.join(rootDir, '.exercism', 'config.json'),
+  ]
+
+  const actualConfigPath = configPaths.find((potentialPath) =>
+    fs.existsSync(potentialPath)
+  )
+
+  if (!actualConfigPath) {
+    return null
+  }
+
+  return JSON.parse(
+    fs.readFileSync(actualConfigPath).toString()
+  ) as ExerciseConfig
 }