+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.1.0
+
+- Upgrade to output version 3
+- Add config.json based handling of task IDs
+- Allow `.exercism` instead of `.meta` for downloaded exercises.
+
 ## 3.0.0
 
 - Rewrite for Jest 27